+++ tests/test_json_schema.cpp
@@ -52,6 +52,8 @@ int main() {
     STAR_CHECK(doc.find("\"Chandra\": \"Bhojana\"") != std::string::npos, "chandra avastha");
     STAR_CHECK(doc.find("\"nakshatra\": \"Puvaputupa\"") != std::string::npos, "details nak");
     STAR_CHECK(doc.find("\"pada\": 2") != std::string::npos, "details pada");
+    STAR_CHECK(doc.find("\"lmt\": \"") != std::string::npos, "times lmt");
+    STAR_CHECK(doc.find("\"gmst\": \"") != std::string::npos, "times gmst");
     STAR_CHECK(doc.find("25\xC2\xB0") != std::string::npos, "details rasi-lon");
     STAR_CHECK(doc.find("\"balance_lord\": \"Guru\"") != std::string::npos, "dasa lord");
     STAR_CHECK(doc.find("\"lord\": \"Guru\", \"from\": \"2000-08-17\"") != std::string::npos,